refactor: move avatar-related elements to common module to be reusable [WPB-20265]
