+++ app/build.gradle.kts
@@ -326,4 +326,6 @@ dependencies {
 
     implementation(project(":ksp"))
     ksp(project(":ksp"))
+
+    testImplementation(testFixtures(project(":core:ui-common")))
 }
