+++ app/src/main/kotlin/com/wire/android/ui/userprofile/self/SelfUserProfileScreen.kt
@@ -54,6 +54,7 @@ import com.ramcosta.composedestinations.result.NavResult
 import com.ramcosta.composedestinations.result.ResultRecipient
 import com.ramcosta.composedestinations.spec.DestinationStyle
 import com.wire.android.R
+import com.wire.android.ui.common.R as UICommonR
 import com.wire.android.appLogger
 import com.wire.android.feature.NavigationSwitchAccountActions
 import com.wire.android.model.ClickBlockParams
@@ -444,10 +445,10 @@ private fun CurrentSelfUserStatus(
         WireDropDown(
             items = items.map {
                 when (it) {
-                    UserAvailabilityStatus.AVAILABLE -> stringResource(R.string.user_profile_status_available)
-                    UserAvailabilityStatus.BUSY -> stringResource(R.string.user_profile_status_busy)
-                    UserAvailabilityStatus.AWAY -> stringResource(R.string.user_profile_status_away)
-                    UserAvailabilityStatus.NONE -> stringResource(R.string.user_profile_status_none)
+                    UserAvailabilityStatus.AVAILABLE -> stringResource(UICommonR.string.user_profile_status_available)
+                    UserAvailabilityStatus.BUSY -> stringResource(UICommonR.string.user_profile_status_busy)
+                    UserAvailabilityStatus.AWAY -> stringResource(UICommonR.string.user_profile_status_away)
+                    UserAvailabilityStatus.NONE -> stringResource(UICommonR.string.user_profile_status_none)
                 }
             },
             defaultItemIndex = items.indexOf(userStatus),
