+++ app/src/main/res/values-ar/strings.xml
@@ -36,7 +36,6 @@
     <string name="label_learn_more">اعرف المزيد</string>
     <string name="label_removing">جارٍ الإزالة…</string>
     <string name="label_logging_in">تسجيل الدخول…</string>
-    <string name="label_membership_guest">ضيف</string>
     <string name="label_user_deleted">محذوف</string>
     <string name="label_message_sent_failure">Message could not be sent.</string>
     <string name="label_message_receive_failure">خطأ في التحميل</string>
