+++ app/src/main/res/values-si/strings.xml
@@ -36,8 +36,6 @@
     <string name="label_learn_more">තව දැනගන්න</string>
     <string name="label_removing">ඉවත් කෙරෙමින්…</string>
     <string name="label_logging_in">පිවිසෙමින්…</string>
-    <string name="label_membership_guest">අමුත්තා</string>
-    <string name="label_membership_external">බාහිර</string>
     <string name="label_user_deleted">මකා ඇත</string>
     <string name="label_try_again">නැවත උත්සාහ කරන්න</string>
     <string name="label_done">කළා</string>
@@ -61,7 +59,6 @@
     </plurals>
     <string name="label_reset_session">වාරය යළි සකසන්න</string>
     <string name="label_reset_session_success">වාරය සාර්ථකව යළි සැකසිණි</string>
-    <string name="label_federated_membership">ඒකාබද්ධ</string>
     <string name="label_update">යාවත්කාල</string>
     <string name="label_close">වසන්න</string>
     <string name="label_download">බාගන්න</string>
@@ -88,7 +85,6 @@
     <string name="content_description_hide_password">මුරපදය සඟවන්න</string>
     <string name="content_description_search_back">ආපසු සොයන්න නිරූපකය</string>
     <string name="content_description_clear_content">අන්තර්ගතය මකන්න</string>
-    <string name="content_description_user_avatar">පැතිකඩ ඡායාරූපය</string>
     <string name="content_description_muted_conversation">නිහඬ කළ සංවාදය</string>
     <string name="content_description_check">හරි ලකුණ</string>
     <string name="content_description_download_icon">බාගැනීමේ නිරූපකය</string>
@@ -598,10 +594,6 @@
     <string name="user_profile_logout">ඉවත් වන්න</string>
     <string name="user_profile_other_accs">ඔබගේ වෙනත් ගිණුම්</string>
     <string name="user_profile_status_availability">තත්‍වය</string>
-    <string name="user_profile_status_available">සිටිනවා</string>
-    <string name="user_profile_status_busy">කාර්යබහුලයි</string>
-    <string name="user_profile_status_away">නොසිටියි</string>
-    <string name="user_profile_status_none">කිසිවක් නැත</string>
     <string name="user_profile_new_account_text">නව කණ්ඩායම හෝ ගිණුමක් එක් කරන්න</string>
     <string name="user_profile_account_management">කණ්ඩායම කළමනාකරණය කරන්න</string>
     <string name="user_profile_details_tab">විස්තර</string>
