+++ core/ui-common/src/main/kotlin/com/wire/android/ui/common/avatar/UnreadInfoIndicator.kt
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2024 Wire Swiss GmbH
+ * Copyright (C) 2025 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,8 +22,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
-import com.wire.android.R
-import com.wire.android.util.ui.PreviewMultipleThemes
+import com.wire.android.ui.common.R
+import com.wire.android.util.PreviewMultipleThemes
 
 @Composable
 fun UnreadInfoIndicator(
