+++ core/ui-common/src/main/kotlin/com/wire/android/util/ui/AssetImageFetcher.kt
