+++ core/ui-common/src/main/res/drawable/ic_blocked_user_avatar.xml
