+++ core/ui-common/src/main/res/drawable/ic_default_service_avatar.xml
