+++ core/ui-common/src/main/res/drawable/ic_default_user_avatar.xml
