+++ core/ui-common/src/main/res/values-ar/strings.xml
@@ -16,4 +16,6 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program. If not, see http://www.gnu.org/licenses/.
   -->
-<resources></resources>
+<resources>
+    <string name="label_membership_guest">ضيف</string>
+</resources>
