+++ core/ui-common/src/main/res/values-fr/strings.xml
@@ -16,4 +16,13 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program. If not, see http://www.gnu.org/licenses/.
   -->
-<resources></resources>
+<resources>
+    <string name="label_membership_guest">Invité.e</string>
+    <string name="label_membership_external">Externe</string>
+    <string name="label_federated_membership">Fédéré</string>
+	<string name="user_profile_status_available">Disponible</string>
+    <string name="user_profile_status_busy">Occupé(e)</string>
+    <string name="user_profile_status_away">Absent(e)</string>
+    <string name="user_profile_status_none">Aucune</string>
+    <string name="content_description_user_avatar">Image de profil</string>
+</resources>
