+++ core/ui-common/src/main/res/values-hr/strings.xml
@@ -16,4 +16,12 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program. If not, see http://www.gnu.org/licenses/.
   -->
-<resources></resources>
+<resources>
+    <string name="label_membership_guest">Gost</string>
+    <string name="label_membership_external">Vanjski</string>
+    <string name="label_federated_membership">Vanjski</string>
+    <string name="user_profile_status_busy">Zauzet</string>
+    <string name="user_profile_status_away">Odsutan</string>
+    <string name="user_profile_status_none">Ništa</string>
+    <string name="content_description_user_avatar">Slika profila</string>
+</resources>
