+++ core/ui-common/src/main/res/values-ja/strings.xml
@@ -16,4 +16,9 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program. If not, see http://www.gnu.org/licenses/.
   -->
-<resources></resources>
+<resources>
+    <string name="label_membership_guest">ゲスト</string>
+    <string name="label_membership_external">外部</string>
+    <string name="label_federated_membership">連合</string>
+    <string name="content_description_user_avatar">プロフィール画像</string>
+</resources>
