+++ core/ui-common/src/main/res/values-pt/strings.xml
@@ -16,4 +16,13 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program. If not, see http://www.gnu.org/licenses/.
   -->
-<resources></resources>
+<resources>
+    <string name="label_membership_guest">Convidado</string>
+    <string name="label_membership_external">Externo</string>
+    <string name="label_federated_membership">Federado</string>
+	<string name="user_profile_status_available">Disponível</string>
+    <string name="user_profile_status_busy">Ocupado</string>
+    <string name="user_profile_status_away">Ausente</string>
+    <string name="user_profile_status_none">Nenhum</string>
+    <string name="content_description_user_avatar">Imagem de perfil</string>
+</resources>
