+++ core/ui-common/src/main/res/values-ru/strings.xml
@@ -33,4 +33,13 @@
     <string name="content_description_alert">Оповещение</string>
     <string name="content_description_show_more_options">Показать больше вариантов</string>
     <string name="content_description_select_label">выбрать</string>
+    <string name="label_membership_guest">Гость</string>
+    <string name="label_membership_external">Внешний</string>
+    <string name="label_membership_service">Приложение</string>
+    <string name="label_federated_membership">Федеративный</string>
+	<string name="user_profile_status_available">Доступен</string>
+    <string name="user_profile_status_busy">Занят</string>
+    <string name="user_profile_status_away">Отошел</string>
+    <string name="user_profile_status_none">Отключен</string>
+    <string name="content_description_user_avatar">Изображение профиля</string>
 </resources>
