+++ core/ui-common/src/main/res/values-si/strings.xml
@@ -16,4 +16,13 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program. If not, see http://www.gnu.org/licenses/.
   -->
-<resources></resources>
+<resources>
+    <string name="label_membership_guest">අමුත්තා</string>
+    <string name="label_membership_external">බාහිර</string>
+    <string name="label_federated_membership">ඒකාබද්ධ</string>
+	<string name="user_profile_status_available">සිටිනවා</string>
+    <string name="user_profile_status_busy">කාර්යබහුලයි</string>
+    <string name="user_profile_status_away">නොසිටියි</string>
+    <string name="user_profile_status_none">කිසිවක් නැත</string>
+    <string name="content_description_user_avatar">පැතිකඩ ඡායාරූපය</string>
+</resources>
