+++ core/ui-common/src/main/res/values-tr/strings.xml
@@ -32,4 +32,8 @@
     <string name="content_description_pending_connection_badge">bağlantı isteğinin onaylanması bekleniyor</string>
     <string name="content_description_alert">Uyarı</string>
     <string name="content_description_show_more_options">Daha fazla seçenek göster</string>
+    <string name="label_membership_guest">Misafir</string>
+    <string name="label_membership_external">Harici</string>
+    <string name="label_federated_membership">Federe</string>
+    <string name="content_description_user_avatar">Profil resmi</string>
 </resources>
