+++ core/ui-common/src/main/res/values-uk/strings.xml
@@ -16,4 +16,9 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program. If not, see http://www.gnu.org/licenses/.
   -->
-<resources></resources>
+<resources>
+    <string name="label_membership_guest">Гість</string>
+    <string name="label_membership_external">Партнер</string>
+    <string name="label_federated_membership">Федеративна система</string>
+    <string name="content_description_user_avatar">Картинка для профілю</string>
+</resources>
