+++ core/ui-common/src/test/kotlin/com/wire/android/util/ui/AssetImageFetcherTest.kt
