+++ gradle.properties
@@ -24,6 +24,7 @@ kotlin.code.style=official
 kotlin.mpp.import.enableKgpDependencyResolution=true
 org.gradle.logging.level=QUIET
 android.experimental.enableScreenshotTest=true
+android.experimental.enableTestFixturesKotlinSupport=true
 
 # Configure lint heap size and workers
 org.gradle.workers.max=4
